vq rolled back

+++ vq.py
@@ -554,7 +554,7 @@ class EuclideanCodebook(nn.Module):
         print(embed_ind)
         # embed_ind = torch.argmax(embed_ind, dim=-1).long()
         # Convert to integer type if needed
-        embed_ind = embed_ind.long()
+        embed_ind = embed_ind.int()
 
         print("embed_ind")
         print(embed_ind)
